🐛 fix: numpy-stable ustrip doctest; drop the None-unsafe hessian annotation
Address PR review:
- The `ustrip(AllowValue, u, AstropyQuantity)` doctest asserted the exact repr
  of a NumPy scalar (`np.float64(1.0)`), which varies across NumPy versions.
  Cast to `float` and assert `1.0` — still exercises the AstropyQuantity
  dispatch, but stable across versions.
- Revert the `hessian` `theunits` annotation to `tuple[AbstractUnit, ...]`. The
  accurate `| None` (matching grad/jacfwd) exposed that the return's
  `theunits[argnums] ** 2` is None-unsafe — `None ** 2` is a hard mypy
  `operator` error, unlike grad/jacfwd's `/ theunits[argnums]`. That return is
  in the unit-correction path already flagged uncertain (`# TODO: check the
  unit correction`), so defining dimensionless-argument semantics there is out
  of scope for this cosmetic PR. Keep the runtime consistency fix (strip with
  the normalized `theunits`, matching grad/jacfwd).

+++ src/unxt/_interop/unxt_interop_astropy/quantity.py
@@ -283,8 +283,8 @@ def ustrip(flag: type[AllowValue], u: Any, x: AstropyQuantity, /) -> Any:
     >>> import astropy.units as apyu
     >>> import unxt as u
     >>> q = apyu.Quantity(1000, "m")
-    >>> u.ustrip(AllowValue, "km", q)
-    np.float64(1.0)
+    >>> float(u.ustrip(AllowValue, "km", q))
+    1.0
 
     """
     return uapi.ustrip(u, x)
